refactor: Separando os slots em arquivos separados
