+++ app/components/ink_components/breadcrumb/list/component.html.erb
@@ -1,0 +1,5 @@
+<%= content_tag :ol, attributes do %>
+  <% items.each do |item| %>
+    <%= item %>
+  <% end %>
+<% end %>
