+++ app/components/ink_components/breadcrumb/list/item/component.html.erb
@@ -1,0 +1,7 @@
+
+<%= content_tag :li, attributes do %>
+  <%= icon %>
+  <%= content_tag :a, href: @href, class: "inline-flex items-center text-sm font-medium text-gray-700 hover:text-pink-600 dark:text-gray-400 dark:hover:text-white" do %>
+    <%= content %>
+  <% end %>
+<% end %>
